+++ src/main/java/com/example/fitpassserver/admin/notice/controller/NoticeAdminController.java
@@ -4,7 +4,7 @@ import com.example.fitpassserver.admin.notice.dto.request.NoticeAdminReqDTO;
 import com.example.fitpassserver.admin.notice.dto.response.AdminNoticeDetailDTO;
 import com.example.fitpassserver.admin.notice.dto.response.NoticeAdminResDTO;
 import com.example.fitpassserver.admin.notice.dto.response.NoticeDraftResDTO;
-import com.example.fitpassserver.admin.notice.service.NoticeAdminService;
+import com.example.fitpassserver.admin.notice.service.NoticeAdminServiceImpl;
 import com.example.fitpassserver.global.apiPayload.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -17,17 +17,16 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 @Tag(name = "Notice 어드민 API", description = "공지사항 관리자 API")
 @RestController
 @RequestMapping("/admin/notice")
 public class NoticeAdminController {
 
-    private final NoticeAdminService noticeAdminService;
+    private final NoticeAdminServiceImpl noticeAdminServiceImpl;
 
-    public NoticeAdminController(NoticeAdminService noticeAdminService) {
-        this.noticeAdminService = noticeAdminService;
+    public NoticeAdminController(NoticeAdminServiceImpl noticeAdminServiceImpl) {
+        this.noticeAdminServiceImpl = noticeAdminServiceImpl;
     }
 
     @Operation(summary = "어드민 공지 목록 조회(키워드 검색 포함)", description = "어드민 공지사항 목록을 조회합니다. 검색어가 없을 땐 파라미터 null")
@@ -38,7 +37,7 @@ public class NoticeAdminController {
             @RequestParam(defaultValue = "10") int size
     ) {
         Pageable pageable = PageRequest.of(page - 1, size);
-        Map<String, Object> noticeList = noticeAdminService.getNoticeAdminList(keyword, pageable);
+        Map<String, Object> noticeList = noticeAdminServiceImpl.getNoticeAdminList(keyword, pageable);
         return ResponseEntity.ok(ApiResponse.onSuccess(noticeList));
     }
 
@@ -48,7 +47,7 @@ public class NoticeAdminController {
             @PathVariable Long noticeId,
             @RequestParam boolean isHomeSlide
     ) {
-        noticeAdminService.updateHomeSlideStatus(noticeId, isHomeSlide);
+        noticeAdminServiceImpl.updateHomeSlideStatus(noticeId, isHomeSlide);
         return ResponseEntity.ok(ApiResponse.onSuccess(null));
     }
 
@@ -58,7 +57,7 @@ public class NoticeAdminController {
             @PathVariable Long noticeId,
             @RequestParam boolean isMemberSlide
     ) {
-        noticeAdminService.updateMemberSlideStatus(noticeId, isMemberSlide);
+        noticeAdminServiceImpl.updateMemberSlideStatus(noticeId, isMemberSlide);
         return ResponseEntity.ok(ApiResponse.onSuccess(null));
     }
 
@@ -69,7 +68,7 @@ public class NoticeAdminController {
             @RequestPart("request") NoticeAdminReqDTO request,
             @RequestPart(value = "image", required = false) MultipartFile image
     ) throws IOException {
-        NoticeAdminResDTO noticeId = noticeAdminService.saveNotice(request, image, true);
+        NoticeAdminResDTO noticeId = noticeAdminServiceImpl.saveNotice(request, image, true);
         return ApiResponse.onSuccess(noticeId);
     }
 
@@ -79,14 +78,14 @@ public class NoticeAdminController {
             @RequestPart("request") NoticeAdminReqDTO request,
             @RequestPart(value = "image",required = false) MultipartFile image
     ) throws IOException {
-        NoticeAdminResDTO response = noticeAdminService.saveNotice(request, image, false);
+        NoticeAdminResDTO response = noticeAdminServiceImpl.saveNotice(request, image, false);
         return ApiResponse.onSuccess(response);
     }
 
     @Operation(summary = "임시 저장된 공지사항 목록 조회")
     @GetMapping(value = "/draftList")
     public ApiResponse<Map<String, Object>> getDraftList() {
-        List<NoticeDraftResDTO> draftNotices = noticeAdminService.getDraftNotices();
+        List<NoticeDraftResDTO> draftNotices = noticeAdminServiceImpl.getDraftNotices();
         Map<String, Object> response = new HashMap<>();
         response.put("notices", draftNotices);
         response.put("count", draftNotices.size());
@@ -95,7 +94,7 @@ public class NoticeAdminController {
     @Operation(summary = "특정 ID의 공지사항 조회")
     @GetMapping("/{id}")
     public ApiResponse<AdminNoticeDetailDTO> getAdminNoticeDetail(@PathVariable Long id) {
-        return ApiResponse.onSuccess(noticeAdminService.getAdminNoticeDetail(id));
+        return ApiResponse.onSuccess(noticeAdminServiceImpl.getAdminNoticeDetail(id));
     }
